(Update) autocorr_cyclo_m
Important note.
There seems to be a bug with FFTW's training of either r2c or many_r2c.
Depending on the training (erase training and restart) r2c or many_r2c don't give the same output (small error 1e-15 but still bigger then it should be). The results should be identical.